More work on smoothing out lines; changed line width from integer to floating point

+++ src/application/plotObject.cpp
@@ -1608,7 +1608,7 @@ void PlotObject::ResetAutoScaling(void)
 //		color		= const Color&
 //		visible		= const bool&
 //		rightAxis	= const bool&
-//		lineSize	= const unsigned int&
+//		lineSize	= const double&
 //		markerSize	= const int&
 //
 // Output Arguments:
@@ -1619,7 +1619,7 @@ void PlotObject::ResetAutoScaling(void)
 //
 //==========================================================================
 void PlotObject::SetCurveProperties(const unsigned int &index, const Color &color,
-	const bool &visible, const bool &rightAxis, const unsigned int &lineSize, const int &markerSize)
+	const bool &visible, const bool &rightAxis, const double &lineSize, const int &markerSize)
 {
 	plotList[index]->SetColor(color);
 	plotList[index]->SetVisibility(visible);
